Add files via upload
implementing  #pragma once  instead of the usual include guards because they looked cool. some general tidying.

+++ Weapon.cpp
@@ -12,7 +12,7 @@ void Weapon::Unequip() {
     attack = 0;
 }
 
-int Weapon::GetAttack() const {
+unsigned short Weapon::GetAttack() const {
     return attack;
 }
